refactor(semconv): Add more unit tests, fix inconsistencies
