+++ crates/weaver_resolver/data/registry-test-lineage-0/expected-registry.json
@@ -30,10 +30,10 @@
       "type": "attribute_group",
       "brief": "Top level group",
       "attributes": [
+        1,
         3,
         4,
-        5,
-        6
+        5
       ],
       "lineage": {
         "source_file": "data/registry-test-lineage-0/registry/groups.yaml",
@@ -42,6 +42,7 @@
             "source_group": "registry.network",
             "inherited_fields": [
               "brief",
+              "examples",
               "note"
             ],
             "locally_overridden_fields": [
@@ -52,6 +53,7 @@
             "source_group": "registry.network",
             "inherited_fields": [
               "brief",
+              "examples",
               "note",
               "requirement_level"
             ]
@@ -60,6 +62,7 @@
             "source_group": "registry.server",
             "inherited_fields": [
               "brief",
+              "examples",
               "note"
             ],
             "locally_overridden_fields": [
