+++ crates/weaver_resolver/data/registry-test-lineage-1/registry/groups.yaml
@@ -14,3 +14,4 @@ groups:
       - id: server.port
         type: string
         brief: Server port
+        examples: ['8080']
