+++ crates/weaver_resolver/data/registry-test-lineage-2/expected-attribute-catalog.json
@@ -3,6 +3,11 @@
     "name": "server.port",
     "type": "string",
     "brief": "Server port",
+    "examples": [
+      80,
+      8080,
+      443
+    ],
     "requirement_level": "recommended",
     "stability": "stable"
   },
@@ -10,18 +15,30 @@
     "name": "network.protocol.name",
     "type": "string",
     "brief": "Network protocol name (registry)",
+    "examples": [
+      "http",
+      "spdy"
+    ],
     "requirement_level": "recommended"
   },
   {
     "name": "network.protocol.version",
     "type": "string",
     "brief": "Network protocol version (registry)",
+    "examples": [
+      "1.0",
+      "2.0"
+    ],
     "requirement_level": "recommended"
   },
   {
     "name": "network.type",
     "type": "string",
     "brief": "Network type (registry)",
+    "examples": [
+      "ipv4",
+      "ipv6"
+    ],
     "requirement_level": "recommended"
   },
   {
@@ -38,6 +55,11 @@
     "name": "server.port",
     "type": "string",
     "brief": "Server port",
+    "examples": [
+      80,
+      8080,
+      443
+    ],
     "requirement_level": "opt_in",
     "stability": "stable"
   },
@@ -55,6 +77,10 @@
     "name": "network.protocol.version",
     "type": "string",
     "brief": "Network protocol version (registry)",
+    "examples": [
+      "1.0",
+      "2.0"
+    ],
     "requirement_level": "opt_in"
   },
   {
