+++ crates/weaver_semconv/data/invalid-semconv.yaml
@@ -1,0 +1,4 @@
+groups:
+  - id: client
+    prefix: client
+    type: attribute_group
