+++ src/registry/stats.rs
@@ -9,7 +9,7 @@ use weaver_common::Logger;
 use weaver_resolved_schema::registry::{CommonGroupStats, GroupStats};
 use weaver_resolved_schema::ResolvedTelemetrySchema;
 use weaver_semconv::group::GroupType;
-use weaver_semconv::SemConvRegistry;
+use weaver_semconv::registry::SemConvRegistry;
 
 /// Parameters for the `registry stats` sub-command
 #[derive(Debug, Args)]
