+++ tests/resolution_process.rs
@@ -3,12 +3,13 @@
 //! Integration tests for the resolution process.
 
 use weaver_cache::Cache;
+use weaver_common::error::ExitIfError;
 use weaver_common::{Logger, TestLogger};
 use weaver_resolver::attribute::AttributeCatalog;
 use weaver_resolver::registry::resolve_semconv_registry;
 use weaver_resolver::SchemaResolver;
 use weaver_semconv::path::RegistryPath;
-use weaver_semconv::SemConvRegistry;
+use weaver_semconv::registry::SemConvRegistry;
 
 /// The URL of the official semantic convention registry.
 const SEMCONV_REGISTRY_URL: &str = "https://github.com/open-telemetry/semantic-conventions.git";
@@ -37,12 +38,8 @@ fn test_semconv_registry_resolution() {
         git_url: SEMCONV_REGISTRY_URL.to_owned(),
         path: Some(SEMCONV_REGISTRY_MODEL.to_owned()),
     };
-    let semconv_specs = SchemaResolver::load_semconv_specs(&registry_path, &cache)
-        .inspect_err(|e| {
-            log.error("Failed to load the semantic convention registry");
-            log.error(&e.to_string());
-        })
-        .unwrap();
+    let semconv_specs =
+        SchemaResolver::load_semconv_specs(&registry_path, &cache).panic_if_error(log.clone());
     let semconv_specs = SemConvRegistry::from_semconv_specs(registry_id, semconv_specs);
 
     // Check if the logger has reported any warnings or errors.
